Update build process to work with new branch config
When this code was developed under the madrailers/sample-app repo it
was configured to detect changes to the master branch and then build
and deploy to the gh-pages branch. Now that we are on the official
organization repo (madrailers.github.io) we need to detect changes to
the source branch and push the built site to the master branch.